backend/requirements.txt
# FastAPI
fastapi==0.115.0
fastapi-mail==1.4.2
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet==3.1.1
databases==0.9.0
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose==3.3.0
python-dotenv==1.0.0
httpx==0.27.2
redis==5.2.0
fastapi-limiter==0.1.6
APScheduler==3.11.0
PyJWT==2.10.1
alembic==1.15.2
cachetools==5.5.2
asyncache==0.3.1
celery==5.5.2
flower==2.0.1

# Twitter API
tweepy==4.15.0

# Web3py
web3==6.19.0

# OAuth
requests==2.32.3
requests-oauthlib>=1.2.0,<2.0.0

passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]